requirements.txt
anthropic
python-dotenv
pymongo
openai
networkx
matplotlib
numpy
pyyaml
sentence-transformers
